+++ handlers/sync_add_to_ratings_handler.go
@@ -27,8 +27,6 @@ func (m SyncAddToRatingsHandler) Handle(options *str.Options, client *internal.C
 		return fmt.Errorf("add to ratings error:%w", err)
 	}
 
-	options.Output = "sync_add_to_ratings_results.json"
-
 	print("write result to:" + options.Output)
 	jsonDataResult, _ := json.MarshalIndent(addResult, "", "  ")
 	writer.WriteJSON(options, jsonDataResult)