+++ handlers/sync_remove_from_favorites_handler.go
@@ -27,8 +27,6 @@ func (m SyncRemoveFromFavoritesHandler) Handle(options *str.Options, client *int
 		return fmt.Errorf("clean favorites error:%w", err)
 	}
 
-	options.Output = "sync_remove_from_favorites_results.json"
-
 	printer.Println("write cleanup result to:" + options.Output)
 	jsonData, _ := json.MarshalIndent(result, "", "  ")
 	writer.WriteJSON(options, jsonData)